+++ frontend/src/components/CharacterSearch.vue
@@ -5,8 +5,7 @@ Input element to search for characters
 <template>
 <div class="input-group input-group-sm mb-1">
     <div class="input-group-prepend">
-        <label class="input-group-text" :class="{ 'search-label-no-padding-left': labelWithoutPaddingLeft }"
-               for="characterSearchInput">
+        <label class="input-group-text" for="characterSearchInput">
             Search {{ admin ? 'Character' : 'Player' }}
         </label>
     </div>
@@ -26,7 +25,6 @@ import {CharacterApi} from 'neucore-js-client';
 export default {
     props: {
         admin: Boolean, // false = search only for mains, otherwise all characters
-        labelWithoutPaddingLeft: Boolean,
     },
 
     data: function() {
@@ -65,9 +63,3 @@ const findCharacter = _.debounce((vm) => {
     }]);
 }, 250);
 </script>
-
-<style scoped>
-    .input-group-text.search-label-no-padding-left {
-        padding-left: 0;
-    }
-</style>